requirements.txt
asttokens==3.0.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
debugpy==1.8.14
decorator==5.2.1
dotenv==0.9.9
executing==2.2.0
Flask==3.1.0
flask-marshmallow==1.3.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
idna==3.10
iniconfig==2.1.0
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.8.1
kafka-python==2.2.11
load-dotenv==0.1.0
MarkupSafe==3.0.2
marshmallow==4.0.0
marshmallow-sqlalchemy==1.4.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pandas-stubs==2.2.3.250308
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.8
pluggy==1.5.0
prompt_toolkit==3.0.51
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
pyzmq==26.4.0
requests==2.32.3
six==1.17.0
spark==0.2.1
SQLAlchemy==2.0.38
stack-data==0.6.3
tornado==6.5.1
traitlets==5.14.3
types-psycopg2==2.9.21.20250516
types-pytz==2025.1.0.20250318
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wcwidth==0.2.13
Werkzeug==3.1.3
